pyproject.toml
[project]
name = "mcp-server-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.1",
    "mcp[cli]>=1.12.4",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
